+++ app/modules/showPath.mjs
@@ -1,0 +1,5 @@
+const ShowPath = () => {
+
+};
+
+export default ShowPath;
